requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0; platform_system != "Windows"  # optional; ignored on some Windows
faiss-cpu>=1.7.4; platform_system != "Windows"
sentence-transformers>=2.7
pypdf>=4.2
python-dotenv>=1.0
requests>=2.32
groq>=0.11.0
tqdm>=4.66
sqlalchemy>=2.0
